+++ src/templates/projectItemTemplate.ts
@@ -1,6 +1,7 @@
 'use strict';
 import * as vscode from 'vscode';
 import * as path from 'path';
+import * as vzFileTemplates from 'vz-file-templates';
 import { ProjectItemTemplateElement } from "./projectItemTemplateElement";
 import { VSCodeHelper } from '../helpers/vscodeHelper';
 import {PathHelper} from '../helpers/pathHelper';
@@ -18,6 +19,9 @@ export class ProjectItemTemplate implements vzFileTemplates.IProjectItemTemplate
     wizardName : string;
     selected : boolean;
     elements : ProjectItemTemplateElement[];
+    isProject : boolean;
+    command : string;
+    commandParameters : string[];
 
     constructor() {
         this.id = 0;
@@ -32,6 +36,9 @@ export class ProjectItemTemplate implements vzFileTemplates.IProjectItemTemplate
         this.templateFilePath = "";
         this.selected = false;
         this.elements = [];
+        this.isProject = false;
+        this.command = "";
+        this.commandParameters = [];
     }
 
     loadFromFile(filePath : string) : boolean {
@@ -56,17 +63,30 @@ export class ProjectItemTemplate implements vzFileTemplates.IProjectItemTemplate
         this.category = content.category as string;
         this.iconLight = content.iconLight as string;
         this.iconDark = content.iconDark as string;
-        this.wizardName = content.wizardName as string;
+        this.wizardName = content.wizardName as string;        
         this.elements = [];
+        this.command = content.command as string;
+        this.commandParameters = [];
+        if (content.isProject)
+            this.isProject = content.isProject as boolean;
 
+        //copy template items
         if (content.elements) {
-            for (let i=0; i<content.elements.length;i++) {
+            for (let i=0; i<content.elements.length; i++) {
                 let templateElement : ProjectItemTemplateElement = new ProjectItemTemplateElement();
                 if (!templateElement.copyFromAny(content.elements[i]))
                     return false;
                 this.elements.push(templateElement);
             }
         }
+
+        //copy command parameters
+        if (content.commandParameters) {
+            for (let i=0; i<content.commandParameters.length; i++) {
+                this.commandParameters.push(content.commandParameters[i] as string);
+            }
+        }
+
         return true;
     }
 
@@ -147,8 +167,76 @@ export class ProjectItemTemplate implements vzFileTemplates.IProjectItemTemplate
             return false;
         }
 
+        //run command
+        if ((settings.command) && (settings.command != "")) {
+            this.runCommand(settings, files);
+        } else {
+            this.runFinished(settings, files);
+        }
+
+        return true;
+    }
+
+    protected runFinished(settings : vzFileTemplates.IProjectItemTemplateRunSettings, files: string[]) {
+        //open workspace
+        if ((!vscode.workspace) || (!vscode.workspace.workspaceFolders) || (vscode.workspace.workspaceFolders.length == 0)) {
+            vscode.commands.executeCommand("vscode.openFolder", vscode.Uri.file(settings.destPath));
+        }
+        //open documents
         for (let i=0; i<files.length; i++) {
             VSCodeHelper.openDocument(vscode.Uri.file(files[i]));
+        }
+    }
+
+    protected runCommand(settings : vzFileTemplates.IProjectItemTemplateRunSettings, files: string[]) : boolean {        
+        let infoText : string = settings.command;
+        
+        //prepare parameters
+        let params : string[] = [];
+        for (let i=0; i<settings.commandParameters.length; i++) {
+            let paramValue : string = settings.applyReplacements(settings.commandParameters[i]);
+            params.push(paramValue);
+            infoText = infoText + " " + paramValue;
+        }
+        
+        //run command
+        try {
+            settings.outputChannel.show();
+            settings.outputChannel.writeLine("Running external command");
+            settings.outputChannel.writeLine(infoText);
+            
+            const childProcess = require('child_process');        
+            const proc = childProcess.spawn(settings.command, params, {
+                cwd: settings.destPath
+            });
+
+            settings.outputChannel.show();
+            settings.outputChannel.writeLine("running command " + settings.command);
+
+            proc.stdout.on('data', (data : any) => {
+                if (data)
+                    settings.outputChannel.write(data.toString());
+            });
+
+            proc.stderr.on('data', (data : any) => {
+                if (data)
+                    settings.outputChannel.write(data.toString());
+            });
+
+            proc.on('error', function(err : string) {
+                vscode.window.showErrorMessage("External command failed with error '" + err + "'.", {modal: false});
+            });
+
+            proc.on('exit', (code : number) => {
+                if (code == 0) {
+                    settings.outputChannel.writeLine("External command completed");
+                    this.runFinished(settings, files);
+                }
+              });
+        }
+        catch (e) {
+            vscode.window.showErrorMessage(e.toString());
+            return false;
         }
 
         return true;